requirements.txt
asgiref==3.6.0
cffi==1.15.1
contextlib2==21.6.0
cryptography==39.0.0
Django==4.1.4
django-ninja==0.20.0
django-ninja-extra==0.18.0
django-ninja-jwt==5.2.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
dnspython==2.2.1
email-validator==1.3.0
idna==3.4
injector==0.19.0
ninja-schema==0.13.0
Pillow==9.4.0
pycparser==2.21
pydantic==1.10.4
PyJWT==2.6.0
pytz==2022.7
sqlparse==0.4.3
typing_extensions==4.4.0
